refactor(openducktor-mcp): split MCP workflow monolith into modules
Separate schema wiring, workflow policy, task mapping, runtime/process helpers, and store context from lib.ts to reduce coupling and make workflow changes safer.
